링크를  분류할 수 있는 섹션 Sort 컴포넌트 구성(기능과 모달창 구현 전)
