+++ W2/react-project/src/index.js
@@ -6,6 +6,6 @@ import "./styles/index.css"
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <App />
+    <App/>
   </React.StrictMode>
 );
