+++ W2/react-project/src/styles/CardList.css
@@ -9,8 +9,6 @@
 
 .card {
   box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-  display: flex;
-  flex-wrap: wrap;
   text-decoration: none;
   height: 20.9rem;
   align-items: flex-start;
@@ -61,4 +59,15 @@
 
 .card:hover .card-image {
   transform: scale(1.1);
+}
+
+.time-createdAt{
+  margin: 0
+}
+.kebab-display{
+  display: flex;
+  justify-content: space-between;
+}
+.kebab-icon{
+  border: none;
 }
